refactor(GameController, InputView): 시도 횟수 입력 로직 수정

+++ src/main/kotlin/racingcar/controller/GameController.kt
@@ -8,7 +8,7 @@ class GameController {
     private val winningCarMembers = mutableListOf<String>()
     fun start() {
         val carMembers = InputView.inputCarName()
-        val numberAttempts = InputView.inputNumberAttempts().toInt()
+        val numberAttempts = InputView.inputNumberAttempts()
         val carList = generateCarList(carMembers)
 
         OutputView.printExecutionResultText()
